fix : put the delete and getsingle in the same file
