+++ src/app/api/posts/[id]/route.js
@@ -1,20 +1,42 @@
 // app/api/posts/[id]/route.ts
 import { NextResponse } from "next/server";
-import { supabase } from "../../../lib/supabase";
-
+import {supabase } from "../../../lib/supabase";
 export async function GET(req, { params }) {
   const { id } = params;
 
   try {
+    // Query the post by its ID
     const { data, error } = await supabase
       .from("posts")
       .select("*")
       .eq("id", id)
-      .single();  // Fetch only one post
+      .single();
 
-    if (error) throw error;
+    if (error) {
+      return NextResponse.json({ error: "Post not found" }, { status: 404 });
+    }
 
-    return NextResponse.json({ post: data });
+    return NextResponse.json(data, { status: 200 });
+  } catch (error) {
+    return NextResponse.json({ error: "Error fetching post" }, { status: 500 });
+  }
+}
+
+export async function DELETE(req, { params }) {
+  const { id } = params;
+
+  try {
+    // Delete the post from the database by its id
+    const { error } = await supabase
+      .from("posts")
+      .delete()
+      .eq("id", id);
+
+    if (error) {
+      throw error;
+    }
+
+    return NextResponse.json({ message: "Post deleted successfully" });
   } catch (error) {
     return NextResponse.json({ error: error.message }, { status: 500 });
   }
